Reimplemented volume annotations

+++ client/src/assets/shaders/annotate.comp.wgsl
@@ -1,8 +1,8 @@
-struct Param 
+struct Param
 {
 	vertex : vec4<f32>,
 	kernelSize : vec4<i32>,
-	clearMask : vec4<i32>,
+	clearMask : vec4<f32>,
 
 	addAnnotation : i32,
   annotationVolume : i32,
